refactor(examples): use reusable workflow components
