+++ examples/custom_backend/user_ensemble/remote/launch.sh
@@ -11,6 +11,7 @@ source "$REPO_ROOT/examples/common/launch_utils.sh"
 trap dynamo_exit_trap EXIT
 
 MODEL="${DYN_MODEL:-Qwen/Qwen2.5-1.5B-Instruct}"
+ENCODER_MODEL="${DYN_ENCODER_MODEL:-$MODEL}"
 ENCODER_CLASS="${DYN_ENCODER_CLASS:-examples.custom_encoder.hitchhikers_vision_encoder.HitchhikersVisionEncoder}"
 CUSTOM_JINJA_TEMPLATE="${DYN_CUSTOM_JINJA_TEMPLATE:-$REPO_ROOT/examples/custom_encoder/templates/qwen_vl.jinja}"
 DECODER_GPU="${DYN_DECODER_GPU:-${CUDA_VISIBLE_DEVICES:-0}}"
@@ -22,6 +23,7 @@ GPU_MEM_ARGS=$(build_vllm_gpu_mem_args)
 
 print_launch_banner --no-curl "Remote User Ensemble" "$MODEL" "$HTTP_PORT" \
     "Encoder:     $ENCODER_CLASS" \
+    "Encoder model: $ENCODER_MODEL" \
     "Decoder GPU: $DECODER_GPU"
 
 export DYN_DISCOVERY_BACKEND="${DYN_DISCOVERY_BACKEND:-file}"
@@ -39,13 +41,14 @@ python3 -m dynamo.frontend \
 
 CUDA_VISIBLE_DEVICES= \
 DYN_SYSTEM_PORT="${DYN_ENCODER_SYSTEM_PORT:-8081}" \
-python3 -m examples.custom_backend.user_ensemble.remote.worker encoder \
-    --model "$MODEL" \
-    --encoder-class "$ENCODER_CLASS" &
+python3 -m dynamo.vllm.workflow.components.encoder_worker \
+    --endpoint-id user-ensemble.encoder.generate \
+    --model "$ENCODER_MODEL" \
+    --custom-encoder-class "$ENCODER_CLASS" &
 
 CUDA_VISIBLE_DEVICES= \
 DYN_SYSTEM_PORT="${DYN_CLASSIFIER_SYSTEM_PORT:-8082}" \
-python3 -m examples.custom_backend.user_ensemble.remote.worker classifier &
+python3 -m examples.custom_backend.user_ensemble.remote.classifier_worker &
 
 CUDA_VISIBLE_DEVICES="$DECODER_GPU" \
 DYN_SYSTEM_PORT="${DYN_DECODER_SYSTEM_PORT:-8083}" \
