+++ examples/custom_backend/user_ensemble/tests/test_workflow.py
@@ -4,7 +4,6 @@
 from __future__ import annotations
 
 from types import SimpleNamespace
-from unittest.mock import AsyncMock
 
 import pytest
 
@@ -21,10 +20,7 @@ pytest.importorskip(
 import torch  # noqa: E402
 
 from dynamo.workflow import ValueRef  # noqa: E402
-from examples.custom_backend.user_ensemble.stages import (  # noqa: E402
-    DummyClassifier,
-    EncoderStage,
-)
+from examples.custom_backend.user_ensemble.stages import DummyClassifier  # noqa: E402
 from examples.custom_backend.user_ensemble.workflow import (  # noqa: E402
     adapt_workflow_result,
     define_workflow,
@@ -66,36 +62,6 @@ def test_workflow_declares_encoder_fanout_to_classifier_and_stock_generator():
     )
     assert generator.inputs["encoder_metadata"] == ValueRef.for_stage_output(
         encoder.id, "encoder_metadata"
-    )
-
-
-async def test_encoder_stage_packs_dynamic_image_rows_and_metadata():
-    first = torch.ones((2, 4), dtype=torch.bfloat16)
-    second = torch.full((3, 4), 2, dtype=torch.bfloat16)
-    encoder = SimpleNamespace(encode=AsyncMock(return_value=[first, second]))
-    stage = EncoderStage(encoder, image_token_id=99)
-    request = {
-        "multi_modal_data": {
-            "image_url": [
-                {"Url": "data:image/png;base64,first"},
-                {"Url": "data:image/png;base64,second"},
-            ]
-        }
-    }
-
-    result = await stage.run({"request": request}, _context())
-
-    assert result["encoder_metadata"] == {
-        "row_splits": [0, 2, 5],
-        "image_token_id": 99,
-    }
-    expected = torch.cat((first, second), dim=0)
-    torch.testing.assert_close(result["encoder_features"], expected)
-    encoder.encode.assert_awaited_once_with(
-        [
-            "data:image/png;base64,first",
-            "data:image/png;base64,second",
-        ]
     )
